# api/docs/swagger.yaml
definitions:
...
  reservation.GetAllReservationOrderRes:
    properties:
      reservation_order:
        items:
          $ref: '#/definitions/reservation.ReservationOrderRes'
        type: array
    type: object
...
  reservation.MenuRes:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
      restaurant:
        $ref: '#/definitions/reservation.Restaurant'
    type: object
...
  reservation.ReservationOrderRes:
    properties:
      id:
        type: string
      menu_item:
        $ref: '#/definitions/reservation.MenuRes'
      quantity:
        type: string
      reservation:
        $ref: '#/definitions/reservation.ReservationRes'
    type: object
  reservation.ReservationOrderUpdate:
    properties:
      menu_item_id:
        type: string
      quantity:
        type: string
      reservation_id:
        type: string
    type: object
...
  reservation.ReservationRes:
    properties:
      id:
        type: string
      reservation_time:
        type: string
      restaurant:
        $ref: '#/definitions/reservation.Restaurant'
      status:
        type: string
      user:
        $ref: '#/definitions/reservation.UserResp'
    type: object
  reservation.Restaurant:
    properties:
      address:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  reservation.UserResp:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /reservation_order:
    post:
      consumes:
      - application/json
      description: Create a new reservation order
      parameters:
      - description: Reservation order data
        in: body
        name: reservation_order
        required: true
        schema:
          $ref: '#/definitions/reservation.ReservationOrderUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.ReservationOrderRes'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create reservation order
      tags:
      - reservation_order
  /reservation_order/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a reservation order by ID
      parameters:
      - description: Reservation order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Reservation order deleted
          schema:
            type: string
        "400":
          description: Invalid reservation order ID
          schema:
            type: string
        "404":
          description: Reservation order not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete reservation order
      tags:
      - reservation_order
    get:
      consumes:
      - application/json
      description: Get a reservation order by ID
      parameters:
      - description: Reservation order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.ReservationOrderRes'
        "400":
          description: Invalid reservation order ID
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get reservation order
      tags:
      - reservation_order
    put:
      consumes:
      - application/json
      description: Update an existing reservation order
      parameters:
      - description: Reservation order ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated reservation order data
        in: body
        name: reservation_order
        required: true
        schema:
          $ref: '#/definitions/reservation.ReservationOrderUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.ReservationOrderUpdate'
        "400":
          description: Invalid request payload
          schema:
            type: string
        "404":
          description: Reservation order not found
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update reservation order
      tags:
      - reservation_order
  /reservation_orders:
    get:
      consumes:
      - application/json
      description: Get all reservation orders
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reservation.GetAllReservationOrderRes'
        "400":
          description: Invalid parameters
          schema:
            type: string
        "500":
          description: Server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all reservation orders
      tags:
      - reservation_order
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"